Feat: Melhoria nas validações do carrinho

+++ src/Services/Carrinho/Carrinho.API/Controllers/CarrinhoController.cs
@@ -6,6 +6,7 @@ using Carrinho.API.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Carrinho.API.Model;
+using FluentValidation.Results;
 using Microsoft.EntityFrameworkCore;
 
 namespace Carrinho.API.Controllers;
@@ -36,6 +37,7 @@ public class CarrinhoController : MainController
             ManipularNovoCarrinho(item);
         else
             ManipularCarrinhoExistente(carrinho, item);
+        ValidarCarrinho(carrinho);
         if (!OperacaoValida()) return CustomResponse();
         await PersistirDados();
         return CustomResponse();
@@ -51,6 +53,8 @@ public class CarrinhoController : MainController
         if (carrinho == null) 
             return CustomResponse();
         carrinho.AtualizarUnidades(itemCarrinho, item.Quantidade);
+        ValidarCarrinho(carrinho);
+        if (!OperacaoValida()) return CustomResponse();
         _context.CarrinhoItens.Update(itemCarrinho);
         _context.CarrinhoClientes.Update(carrinho);
         await PersistirDados();
@@ -66,6 +70,8 @@ public class CarrinhoController : MainController
             return CustomResponse();
         if (carrinho == null) 
             return CustomResponse();
+        ValidarCarrinho(carrinho);
+        if (!OperacaoValida()) return CustomResponse();
         carrinho.RemoverItem(itemCarrinho);
         _context.CarrinhoItens.Remove(itemCarrinho);
         _context.CarrinhoClientes.Update(carrinho);
@@ -127,5 +133,13 @@ public class CarrinhoController : MainController
         var result = await _context.SaveChangesAsync();
         if (result<=0)
             AdicionarErroProcessamento("Não foi possível persistir os dados no banco.");
+    }
+
+    private bool ValidarCarrinho(CarrinhoCliente carrinho)
+    {
+        if (carrinho.EhValido())
+            return true;
+        carrinho.ValidationResult.Errors.ToList().ForEach(e => AdicionarErroProcessamento(e.ErrorMessage));
+        return false;
     }
 }
